+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ## Master
 
+- Add commit ID to the comment footer [danger/danger-js#168] - [@randak][]
 - Add support for `COPY` change type to fix a BitBucket Server regression in
   [danger/danger-js#764](https://github.com/danger/danger-js/pull/764) - [@sebinsua][]
 - Add support for older Babel versions (prior 7) [@sajjadzamani][]
@@ -1435,4 +1436,5 @@ Not usable for others, only stubs of classes etc. - [@orta][]
 [@patrickkempff]: https://github.com/patrickkempff
 [@imorente]: https://github.com/imorente
 [@langovoi]: https://github.com/langovoi
+[@randak]: https://github.com/randak
 [@sajjadzamani]: https://github.com/sajjadzamani